sondia/pyproject.toml
[project]
name = "sondia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.5.3",
    "fastapi>=0.115.14",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
